Set up the general workflow of the optimizer

+++ Tests/TestOptimization.cs
@@ -13,11 +13,11 @@ public class TestOptimization
     [Test]
     public void TestWD()
     {
-        var a = new List<double>() { 0, 1, 3 };
-        var b = new List<double>() { 1, 2, 1 };
-        var aCDF = StatisticMeasures<double>.GetCDF(a);
+        var a = new List<int>() { 0, 1, 3 };
+        var b = new List<int>() { 1, 2, 1 };
+        var aCDF = StatisticMeasures<int>.GetCDF(a);
         Assert.AreEqual(.25, aCDF[1], double.Epsilon);
         Assert.AreEqual(1, aCDF[2], double.Epsilon);
-        var dist = StatisticMeasures<double>.WassersteinDistance(a, b);
+        var dist = StatisticMeasures<int>.WassersteinDistance(a, b);
     }
 }
